# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Luma Sticker Store API",
    "description": "Dokumentasi API untuk project Luma Sticker (E-commerce Digital Assets).",
    "version": "1.0.0",
    "contact": {
      "name": "Abdian Developer",
      "email": "abdian@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local Server (Port 3000)"
    },
    {
      "url": "https://lumasticker.vercel.app/api",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Sticker Pack Vol. 1 (Chibi Style)"
          },
          "price": {
            "type": "number",
            "example": 15000
          },
          "description": {
            "type": "string",
            "example": "Stiker lucu format PNG & PDF resolusi tinggi."
          },
          "image_url": {
            "type": "string",
            "description": "URL gambar utama (diambil dari images[0])",
            "example": "http://localhost:5000/uploads/1700000000-file.png"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "filename1.png",
              "filename2.png"
            ]
          },
          "is_active": {
            "type": "integer",
            "example": 1,
            "description": "1 = Active/Published, 0 = Inactive/Draft"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Ambil Semua Produk (Public)",
        "responses": {
          "200": {
            "description": "List Produk",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Tambah Produk Baru (Admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "price",
                  "description"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "description": {
                    "type": "string"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Upload hingga 20 gambar"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produk Created"
          },
          "401": {
            "description": "Unauthorized (Token Kosong/Salah)"
          },
          "403": {
            "description": "Forbidden (Token Invalid)"
          }
        }
      }
    },
...
    "/products/{id}": {
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update Produk",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "description": {
                    "type": "string"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update Berhasil"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Hapus Produk Satuan",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Terhapus"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}